% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HandwrittenDigits.R
\name{HandwrittenDigits}
\alias{HandwrittenDigits}
\title{Handwritten Digit Recognition based on kNN classification.}
\usage{
HandwrittenDigits(data, trueClasses = NULL, memory = NULL, k = 1, p = 2,
  type = "train")
}
\arguments{
\item{data}{A data frame or a matrix where rows are observations and columns are features. If \code{type} is "train" this is training dataset, and if it is "predict" it is test dataset.}

\item{trueClasses}{A vector with labels for each row in \code{data} if \code{type} is "train", and with labels for each row in \code{memory} if \code{type} is "predict".}

\item{memory}{A data frame or a matrix where rows are observations and columns are features. If \code{type} is "train" this argument is not needed, and if it is "predict" it is a training dataset.}

\item{k}{Number of neighbors that the classifier should use. It has to be an odd number.}

\item{p}{Distance metric the classifier should use, the value can be either 1, 2 or Inf.}

\item{type}{Whether the goal is to train the classifier or predict classes of new observations based on past ones. The value can be either "train" or "predict".}
}
\value{
A list with following elements: predictedClasses, prob, accuracy and errorCount.
}
\description{
Classify the input using a kNN classification.
}
\examples{
# create artificial dataset
inputsTrain  <- cbind(sample(0:9, 500, replace=TRUE), matrix(rep(runif(256,min = -1, max = 1), 500), 500))
inputsTest  <- matrix(rep(runif(256,min = -1, max = 1), 200), 200)
classesTrain <- inputsTrain[,1]

# get the kNN predictions for the train set
trainResults <- HandwrittenDigits(inputsTrain[,2:257], classesTrain, k=15, p=2, type="train")
# get the kNN predictions for the test set
testResults <- HandwrittenDigits(data=inputsTest[,1:256], trueClasses=classesTrain, memory=inputsTrain[,2:257], k=15, p=2, type="predict")
}

